Improving Proc() function
This function first finds the <textarea> element where the song code is written.
If it doesn't exist, it stops and prevents crash.

const raw = textArea.value || "";
Reads the original text  (whatever is in the text area) and then if the value is undefined or empty it uses an empty string.

It also Checks if the hush radio button is selected. ?. handles cases where the element doesn't exist. Followed with a conditional assignment with strudel syntax criteria.

const replaced finds every <p1_radio> tag in the user's text and replaces it with correct value.