Improving SetupButtons() functions and adding a tempo feature
Simply find the button in the HTML and give each one a specific job when you click it for example, play music, stop music, or preprocess code.

Implementing a tempo feature, which will allow me to change the playback speed of my strudel directly from the input box in the UI. Allowing me to have dynamic control over the BPM of the song.

This code includes safety checks in case of any failures.